refactor: Modify the data types and foreign keys in ORM models

+++ modules/imdb/src/main/java/dev/sonle/filmdb/imdb/dto/MovieFilterSortRequestDto.java
@@ -1,10 +1,8 @@
 package dev.sonle.filmdb.imdb.dto;
 
-import java.math.BigDecimal;
-
 public record MovieFilterSortRequestDto(
     Integer startYear,
-    BigDecimal averageRating,
+    Float averageRating,
     Integer numVotes,
     String titleType,
     String sortBy,
